[Chenyu Zhao] : optimize logic directory structure
